fix: change ENUM into new table, functional
search by job not functional

+++ src/main/java/overlook_hotel/overlook_hotel/repository/EmployeeRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import overlook_hotel.overlook_hotel.model.entity.Employee;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer>, JpaSpecificationExecutor<Employee> {
-    // Use DatabaseEnumService for enum fetching instead of a hardcoded query here.
+public interface EmployeeRepository extends JpaRepository<Employee, Long>, JpaSpecificationExecutor<Employee> {
 }
